+++ src/main/java/cn/itdeer/kafka/core/type/NoAppointJsonProducer.java
@@ -31,7 +31,7 @@ public class NoAppointJsonProducer extends Thread {
      * 基本信息
      */
     private String topicName;
-    private int dataNumber;
+    private long dataNumber;
     private int timeFrequency;
 
     /**
@@ -57,7 +57,7 @@ public class NoAppointJsonProducer extends Thread {
     public void run() {
         long startTime = System.currentTimeMillis();
         String startDate = Constants.format.format(new Date());
-        int totleNumber = dataNumber;
+        long totleNumber = dataNumber;
 
         /**
          * 初始化
